--- source/class/qx/tool/cli/CompileJs.js
@@ -1,412 +1,0 @@
-/* ************************************************************************
-
-   qooxdoo - the new era of web development
-
-   http://qooxdoo.org
-
-   Copyright:
-     2017 Zenesis Ltd
-
-   License:
-     MIT: https://opensource.org/licenses/MIT
-     See the LICENSE file in the project's top-level directory for details.
-
-   Authors:
-     * John Spackman (john.spackman@zenesis.com, @johnspackman)
-     * Henner Kollmann(Henner.Kollmann@gmx.de, @hkollmann)
-
-************************************************************************ */
-
-
-require("@qooxdoo/framework");
-require("../config");
-const path = require("upath");
-const fs = qx.tool.utils.Promisify.fs;
-const semver = require("semver");
-
-
-/**
- * Mixin used for commands that need to be able to parse compiler configuration
- * 
- * Configuration files do not support processes, job executions, or even
- * macros - if you want to add basic processing (eg for macros), use a .js file to
- * manipulate the data.  If you want to customise the Maker that is produced, you
- * need to use the API directly.
- *
- */
-qx.Class.define("qx.tool.cli.CompileJs", {
-  extend: qx.core.Object,
-  
-  construct(argv) {
-    this.base(arguments);
-    this.argv = argv;
-  },
-  
-  statics: {
-    compileJsFilename: "compile.js",
-    
-    /** {CompileJs} singleton instance */
-    __instance: null,
-    
-    /**
-     * Returns the singleton instance, throws an error if it has not been created
-     * 
-     * @return {CompileJs}
-     */
-    getInstance() {
-      const CompileJs = qx.tool.cli.CompileJs;
-      if (!CompileJs.__instance) {
-        throw new Error("CompileJs has not been initialized yet!");
-      }
-      return CompileJs.__instance;
-    },
-    
-    /**
-     * Sets the singleton, can only be done once
-     * 
-     * @param instance {CompileJs}
-     */
-    setInstance(instance) {
-      const CompileJs = qx.tool.cli.CompileJs;
-      if (CompileJs.__instance) {
-        throw new Error("CompileJs has already been initialized!");
-      }
-      CompileJs.__instance = instance;
-    }
-  },
-
-  members: {
-    _compilerApi: null,
-    _compileJsFilename: null,
-    _compileJsonFilename: null,
-    _parsedArgs: null,
-    
-    /**
-     * Parses the command line and produces configuration data; this is safe to
-     * call multiple times, provided that you await each time you call it.
-     * 
-     * Loads a configuration file from a .js or .json file; if you provide a .js
-     * file the file must be a module which returns an object which has any of
-     * these properties:
-     * 
-     *  CompilerConfig - the class (derived from qx.tool.cli.api.CompilerApi)
-     *    for configuring the compiler
-     *    
-     * Each library can also have a compile.js, and that is also a module which can
-     * return an object with any of these properties:
-     *  
-     *  LibraryConfig - the class (derived from qx.tool.cli.api.LibraryApi)
-     *    for configuring the library 
-     *
-     */
-    parse: async function() {
-      if (this._parsedArgs) {
-        return this._parsedArgs;
-      }
-      
-      let parsedArgs = {
-        target: this.argv.target,
-        outputPath: null,
-        locales: null,
-        writeAllTranslations: this.argv.writeAllTranslations,
-        environment: {},
-        config: this.argv.configFile||qx.tool.config.Compile.config.fileName,
-        verbose: this.argv.verbose
-      };
-
-      if (this.argv.locale && this.argv.locale.length) {
-        parsedArgs.locales = this.argv.locale;
-      }
-
-      if (this.argv["set-env"]) {
-        this.argv["set-env"].forEach(function(kv) {
-          var m = kv.match(/^([^=\s]+)(=(.+))?$/);
-          var key = m[1];
-          var value = m[3];
-          parsedArgs.environment[key] = value;
-        });
-      }
-      
-      var lockfile_content = {
-        version: qx.tool.config.Lockfile.getInstance().getVersion()
-      };
-
-      let compileJsFilename = qx.tool.cli.CompileJs.compileJsFilename;
-      let compileJsonFilename = qx.tool.config.Compile.config.fileName;
-      if (parsedArgs.config) {
-        if (parsedArgs.config.match(/\.js$/)) {
-          compileJsFilename = parsedArgs.config;
-        } else {
-          compileJsonFilename = parsedArgs.config;
-        }
-      }
-      
-      if (await fs.existsAsync(compileJsonFilename)) {
-        this._compileJsonFilename = compileJsonFilename;
-      }
-      
-      let CompilerApi = qx.tool.cli.api.CompilerApi;
-      if (await fs.existsAsync(compileJsFilename)) {
-        let compileJs = await this.__loadJs(compileJsFilename);
-        this._compileJsFilename = compileJsFilename;
-        if (compileJs.CompilerApi) {
-          CompilerApi = compileJs.CompilerApi;
-        }
-      }
-      let compilerApi = this._compilerApi = new CompilerApi(this).set({ 
-        rootDir: ".",
-        configFilename: compileJsonFilename
-      });
-      
-      await compilerApi.load();
-      let config = compilerApi.getConfiguration();
-      
-      if (parsedArgs.config) {
-        let lockfile = qx.tool.config.Lockfile.config.fileName;
-        try {
-          var name = path.join(path.dirname(parsedArgs.config), lockfile);
-          lockfile_content = await qx.tool.utils.Json.loadJsonAsync(name) || lockfile_content;
-        } catch (ex) {
-          // Nothing
-        }
-        // check semver-type compatibility (i.e. compatible as long as major version stays the same)
-        let schemaVersion = semver.coerce(qx.tool.config.Lockfile.getInstance().getVersion(), true).raw;
-        let fileVersion = lockfile_content && lockfile_content.version ? semver.coerce(lockfile_content.version, true).raw : "1.0.0";
-        if (semver.major(schemaVersion) > semver.major(fileVersion)) {
-          if (this.argv.force) {
-            let config = {
-              verbose: this.argv.verbose,
-              quiet: this.argv.quiet,
-              save: false
-            };
-            const installer = new qx.tool.cli.commands.package.Install(config);
-            let filepath = installer.getLockfilePath();
-            let backup = filepath + ".old";
-            await fs.copyFileAsync(filepath, backup);
-            if (!this.argv.quiet) {
-              qx.tool.compiler.Console.warn(`*** A backup of ${lockfile} has been saved to ${backup}, in case you need to revert to it. ***`);
-            }
-            await installer.deleteLockfile();
-            for (let lib of lockfile_content.libraries) {
-              if (!await installer.isInstalled(lib.uri, lib.repo_tag)) {
-                if (lib.repo_tag) {
-                  await installer.install(lib.uri, lib.repo_tag);
-                } else if (lib.path && fs.existsSync(lib.path)) {
-                  await installer.installFromLocaPath(lib.path, lib.uri);
-                }
-              } else if (this.argv.verbose) {
-                qx.tool.compiler.Console.info(`>>> ${lib.uri}@${lib.repo_tag} is already installed.`);
-              }
-            }
-            lockfile_content = await installer.getLockfileData();
-          } else {
-            throw new qx.tool.utils.Utils.UserError(
-              `*** Warning ***\n` +
-              `The schema of '${lockfile}' has changed. Execute 'qx clean && qx compile --force' to delete and regenerate it.\n` +
-              `You might have to re-apply manual modifications to '${lockfile}'.`
-            );
-          }
-        }
-      }
-      this._mergeArguments(parsedArgs, config, lockfile_content);
-
-      if (config.libraries) {
-        /*
-         * Locate and load libraries
-         */
-        if (!config.libraries.every(libData => fs.existsSync(libData + "/Manifest.json"))) {
-          qx.tool.compiler.Console.log("One or more libraries not found - trying to install them from library repository...");
-          const installer = new qx.tool.cli.commands.package.Install({
-            quiet: true,
-            save: false
-          });
-          await installer.process();
-        }
-
-        for (const aPath of config.libraries) {
-          let libCompileJsFilename = path.join(aPath, qx.tool.cli.CompileJs.compileJsFilename);
-          let LibraryApi = qx.tool.cli.api.LibraryApi;
-          if (await fs.existsAsync(libCompileJsFilename)) {
-            let compileJs = await this.__loadJs(libCompileJsFilename);
-            if (compileJs.LibraryApi) {
-              LibraryApi = compileJs.LibraryApi;
-            }
-          }
-          
-          let libraryApi = new LibraryApi().set({ 
-            rootDir: aPath,
-            compilerApi: compilerApi
-          });
-          compilerApi.addLibraryApi(libraryApi);
-          await libraryApi.load();
-        }
-      }
-      
-      if (config.sass && config.sass.compiler !== undefined) {
-        qx.tool.compiler.resources.ScssConverter.USE_V6_COMPILER = config.sass.compiler == "latest";
-      } else {
-        qx.tool.compiler.resources.ScssConverter.USE_V6_COMPILER = null;
-      }
-      
-      await compilerApi.afterLibrariesLoaded();
-      
-      this._parsedArgs = await compilerApi.getConfiguration();
-      return this._parsedArgs;
-    },
-
-    /*
-     * Merges the argv into the config data
-     *
-     */
-    _mergeArguments: function(parsedArgs, config, lockfileContent) {
-      config.targetType = parsedArgs.target||config.defaultTarget||"source";
-
-      if (!config.locales) {
-        config.locales = [];
-      }
-      if (typeof parsedArgs.writeAllTranslations == "boolean") {
-        config.writeAllTranslations = parsedArgs.writeAllTranslations;
-      }
-
-      if (!config.environment) {
-        config.environment = {};
-      }
-
-      // Set the environment variables coming from command line arguments
-      // in target's environment object. If that object doesn't exist create
-      // one and assign it to the target.
-      if (config.targets) {
-        const target = config.targets.find(target =>
-          target.type === config.targetType);
-        target.environment = target.environment || {};
-        qx.lang.Object.mergeWith(target.environment, parsedArgs.environment, true);
-      }
-
-      if (!config.libraries) {
-        if (fs.existsSync("Manifest.json")) {
-          config.libraries = [ "." ];
-        }
-      }
-
-      if (lockfileContent.libraries) {
-        config.packages = {};
-        lockfileContent.libraries.forEach(function(library) {
-          config.libraries.push(library.path);
-          config.packages[library.uri] = library.path;
-        });
-      }
-
-      if (!config.serve) {
-        config.serve = {};
-      }
-      if (this.isExplicitArg("listen-port")) {
-        config.serve.listenPort = this.argv.listenPort;
-      } else {
-        config.serve.listenPort = config.serve.listenPort || this.argv.listenPort;
-      }
-    },
-
-    /**
-     * Parses the command line, and produces a normalised configuration;
-     *
-     * @return {Object}
-     */
-    __parseImpl: async function() {
-      let argv = this.argv;
-      let result = {
-        target: argv.target,
-        outputPath: null,
-        locales: null,
-        writeAllTranslations: argv.writeAllTranslations,
-        environment: {},
-        config: argv.configFile||qx.tool.config.Compile.config.fileName,
-        verbose: argv.verbose
-      };
-
-      if (argv.locale && argv.locale.length) {
-        result.locales = argv.locale;
-      }
-
-      if (argv["set-env"]) {
-        argv["set-env"].forEach(function(kv) {
-          var m = kv.match(/^([^=\s]+)(=(.+))?$/);
-          var key = m[1];
-          var value = m[3];
-          result.environment[key] = value;
-        });
-      }
-
-      return result;
-    },
-
-    /**
-     * Loads a .js file using `require`, handling exceptions as best as possible
-     * 
-     * @param aPath {String} the file to load
-     * @return {Object} the module
-     */
-    __loadJs: async function(aPath) {
-      try {
-        let module = require(path.resolve(aPath));
-        return module;
-      } catch (e) {
-        let lines = e.stack.split("\n");
-        for (let i = 0; i < lines.length; i++) {
-          if (lines[i].match(/^\s+at/)) {
-            lines.splice(i); 
-          }
-        }
-        let lineNumber = lines[0].split("evalmachine.<anonymous>:")[1];
-        if (lineNumber !== undefined) {
-          lines.shift();
-          throw new Error("Error while reading " + aPath + " at line " + lineNumber + "\n" + lines.join("\n"));
-        } else {
-          throw new Error("Error while reading " + aPath + "\n" + lines.join("\n"));
-        }
-      }
-    },
-    
-    /**
-     * Returns the CompilerApi instance
-     * 
-     * @return {CompilerApi}
-     */
-    getCompilerApi() {
-      return this._compilerApi;
-    },
-    
-    /**
-     * Returns the filename of compile.js, if there is one
-     * 
-     * @return {String?} filename
-     */
-    getCompileJsFilename() {
-      return this._compileJsFilename;
-    },
-
-    /**
-     * Returns the filename of compile.json, if there is one
-     * 
-     * @return {String?} filename
-     */
-    getCompileJsonFilename() {
-      return this._compileJsonFilename;
-    },
-    
-    /**
-     * Detects whether the command line explicit set an option (as opposed to yargs
-     * providing a default value).  Note that this does not handle aliases, use the
-     * actual, full option name.
-     *
-     * @param option {String} the name of the option, eg "listen-port"
-     * @return {Boolean}
-     */
-    isExplicitArg(option) {
-      function searchForOption(option) {
-        return process.argv.indexOf(option) > -1;
-      }
-      return searchForOption(`-${option}`) || searchForOption(`--${option}`);
-    }
-
-  }
-});
