+++ source/class/qx/tool/cli/index.js
@@ -43,6 +43,3 @@ require("./commands/package/Migrate.js");
 require("./api/AbstractApi.js");
 require("./api/CompilerApi.js");
 require("./api/LibraryApi.js");
-
-require("./CompileJs.js");
-
